+++ src/base/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(base_SRCS
     Timestamp.cpp
+    PThreadMutex.cpp
 )
 
 add_library(cppevt_base ${base_SRCS})
