+++ src/base/PThreadMutex.h
@@ -1,7 +1,7 @@
 #ifndef __MUTEX_H__
 #define __MUTEX_H__
 
-#include <net/NonCopyable.h>
+#include <base/NonCopyable.h>
 
 #include <pthread.h>
